requirements.txt
absl-py==1.4.0
addict==2.2.1
albumentations==1.3.0
antspyx==0.2.2
anytree==2.7.3
appdirs==1.4.4
async-timeout==4.0.2
atomicwrites==1.3.0
attrs==19.3.0
batchgenerators==0.21
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
chardet==2.3.0
charset-normalizer==2.0.12
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
colorlog==6.7.0
coverage==5.5
cycler==0.10.0
Cython==0.29.21
dask==2.6.0
dataclasses==0.8
DBUtils==1.3
decorator==4.4.1
defer==1.0.4
dicom2nifti==2.1.8
dijkstra3d==1.1.0
dipy==1.0.0
docopt==0.6.2
dwclustersdk==0.4.0
easydict==1.10
edt==2.0.1
einops==0.4.1
et-xmlfile==1.1.0
fastremap==1.12.2
filelock==3.0.12
flatbuffers==1.11
funcsigs==1.0.2
future==0.18.2
fvcore==0.1.3.post20210204
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.48.2
h5py==2.10.0
hdfs==2.6.0
html2text==2020.1.16
idna==3.4
imagecorruptions==0.1.0
imageio==2.6.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
inplace-abn==1.1.0
iopath==0.1.8
itk==5.2.1.post1
itk-core==5.2.1.post1
itk-elastix==0.13.0
itk-filtering==5.2.1.post1
itk-io==5.2.1.post1
itk-numerics==5.2.1.post1
itk-registration==5.2.1.post1
itk-segmentation==5.2.1.post1
joblib==1.1.1
json-tricks==3.16.1
json5==0.9.11
jsonschema==3.2.0
kiwisolver==1.3.1
kornia==0.5.0
linecache2==1.0.0
llvmlite==0.34.0
lmdb==1.4.1
lxml==4.9.2
Mako==1.1.6
Markdown==3.3.7
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528150516/work
matplotlib==3.3.4
MedPy==0.4.0
meshio==4.1.0
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mmcv-full==1.6.1
mmdet==2.25.3
mmengine==0.4.0
monai==0.5.3
more-itertools==7.2.0
mpmath==1.3.0
msgpack==1.0.5
mysqlclient==1.4.4
networkx==2.4
nibabel==2.5.1
nose==1.3.7
numba==0.51.2
numpy==1.19.5
oauthlib==3.2.2
onnxruntime==1.10.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
packaging==21.3
pandas==0.25.2
patsy==0.5.3
pexpect==4.0.1
pickle5==0.0.10
Pillow==6.2.1
platformdirs @ file:///tmp/build/80754af9/platformdirs_1638968569495/work
pluggy==0.13.0
portalocker==2.3.0
protobuf==3.19.6
psutil==3.2.2
ptyprocess==0.5
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycollada==0.7.2
pycrypto==2.6.1
pycuda @ file:///home/conda/feedstock_root/build_artifacts/pycuda_1617557199483/work
pycurl==7.43.0
pydensecrf==1.0rc2
pydicom==2.3.1
pyefd==1.6.0
pyglet==2.0.5
pykwalify==1.8.0
pymeshfix==0.13.4
pynrrd==0.4.3
pynvml==8.0.3
pyparsing==2.4.2
pyradiomics==3.0.1
pyrsistent==0.18.0
pytest==5.2.2
pytest-runner==5.3.0
python-dateutil==2.8.0
python-systemd==0.0.9
pytools @ file:///home/conda/feedstock_root/build_artifacts/pytools_1656247651513/work
pytz==2019.3
pyvista==0.24.2
PyWavelets==1.1.1
PyYAML==3.12
qudida==0.0.4
ray==0.6.6
redis==3.3.11
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
Rtree==0.9.7
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.16.2
scikit-learn==0.20.2
scikit-spatial==5.2.0
scipy==1.3.1
scooby==0.5.6
setproctitle==1.1.10
Shapely==1.8.5.post1
SimpleITK==2.1.1.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
ssh-import-id==5.5
statsmodels==0.12.2
svg.path==6.2
sympy==1.9
tabulate==0.8.9
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorrt @ file:///mnt/users/package/TensorRT-8.0.1.6/python/tensorrt-8.0.1.6-cp36-none-linux_x86_64.whl
termcolor==1.1.0
terminaltables==3.1.0
tetgen==0.4.4
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.4.12
tomli==1.2.3
toolz==0.10.0
torch @ https://download.pytorch.org/whl/cu113/torch-1.10.0%2Bcu113-cp36-cp36m-linux_x86_64.whl
torchinfo==1.5.4
torchvision @ https://download.pytorch.org/whl/cu113/torchvision-0.11.0%2Bcu113-cp36-cp36m-linux_x86_64.whl
tornado==6.1
tqdm==4.36.1
traceback2==1.4.0
triangle==20220202
trimesh==3.9.35
typing==3.7.4.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
unittest2==1.1.0
urllib3==1.26.7
vtk==8.1.2
wcwidth==0.1.7
webcolors==1.11.1
Werkzeug==2.0.3
xlrd==1.1.0
xxhash==3.2.0
yacs==0.1.6
yapf==0.30.0
zipp==0.6.0
